finishing pinecone

+++ 02- Pinecone-Image/main.py
@@ -18,17 +18,19 @@ os.makedirs(DOWNLOADED_IMAGES_FOLDER_PATH, exist_ok=True)
 
 ## --------------------------------- EndPoint for Getting Similar Images ------------------------------ ##
 @app.post('/image-search')
-async def image_search(image_url: str=Form(...), top_k: int=Form(...), threshold: float=Form(0.5)):
+async def image_search(image_url: str=Form(...), top_k: int=Form(...), threshold: float=Form(None), 
+                       class_type: str=Form(..., description='class_type', enum=['All', 'class-a', 'class-b'])):
 
     ## Validate top_k and threshold
     if top_k <= 0 or not isinstance(top_k, int) or top_k > 10000:
         raise HTTPException(status_code=400, detail="Bad Request: 'top_k' must be between integer and between 1 and 10000.")
 
-    elif threshold <= 0.0 or not isinstance(threshold, float) or threshold > 1.0:
+    elif threshold is not None and (threshold <= 0.0 or not isinstance(threshold, float) or threshold > 1.0):
         raise HTTPException(status_code=400, detail="Bad Request: 'threshold' must be between 0 and 1.")
 
     ## Call (search_vectorDB) from utils.py --> for searching
-    results = search_vectorDB(image_url=image_url, folder_path=DOWNLOADED_IMAGES_FOLDER_PATH, top_k=top_k, threshold=threshold)
+    results = search_vectorDB(image_url=image_url, folder_path=DOWNLOADED_IMAGES_FOLDER_PATH, 
+                              top_k=top_k, threshold=threshold, class_type=class_type)
 
     ## After Getting the response -- Remove the images
     _ = [os.remove(os.path.join(DOWNLOADED_IMAGES_FOLDER_PATH, image)) for image in os.listdir(DOWNLOADED_IMAGES_FOLDER_PATH)]
@@ -48,15 +50,17 @@ os.makedirs(INDEXING_IMAGES_FOLDER_PATH, exist_ok=True)
 ## ---------------------------------------------- EndPoint for Updates ------------------------------------------- ##
 @app.post('/indexing_or_deleting')
 async def indexing_or_deleting(image_id: int=Form(...), image_url: str=Form(None), 
+                               class_type: str=Form(None, description='class_type', enum=['class-a', 'class-b']),
                                case: str=Form(..., description='case', enum=['upsert', 'delete'])):
 
     ## Validate the new_text is not None if the case=upsert
-    if case == 'upsert' and not image_url:
-        raise HTTPException(status_code=400, detail='"image_url" is mandatory for case "upsert".')
+    if case == 'upsert' and (not image_url or not class_type):
+        raise HTTPException(status_code=400, detail='"image_url & class_type" is mandatory for case "upsert".')
     
     ## For Upserting
     if case == 'upsert':
-        message = insert_vectorDB(image_id=image_id, image_url=image_url, folder_path=INDEXING_IMAGES_FOLDER_PATH)     
+        message = insert_vectorDB(image_id=image_id, image_url=image_url, 
+                                  folder_path=INDEXING_IMAGES_FOLDER_PATH, class_type=class_type)     
         ## After Getting the response -- Remove the images
         _ = [os.remove(os.path.join(INDEXING_IMAGES_FOLDER_PATH, image)) for image in os.listdir(INDEXING_IMAGES_FOLDER_PATH)]
 
@@ -64,5 +68,4 @@ async def indexing_or_deleting(image_id: int=Form(...), image_url: str=Form(None
     elif case == 'delete':
         message = delete_vectorDB(image_id=image_id)
 
-    return {'message': message}                           
-
+    return {'message': message}
